+++ src/command.rs
@@ -150,11 +150,13 @@ fn shop(game: &mut Game, items: &[String]) -> Result<()> {
     if items.is_empty() {
         item::shop::list(game)
     } else {
-        for item_name in items {
-            let item_name = Key::from(item_name)?;
-            item::shop::buy(game, item_name)?
+        // parse items and break if any is invalid/unknown
+        let mut keys = Vec::new();
+        for item in items {
+            keys.push(Key::from(item)?);
         }
-        Ok(())
+
+        item::shop::buy(game, &keys)
     }
 }